Speed up JDJsonDecoder
Use C functions instead of Objective-C messages
Removed attribute calculation not required

+++ JDJsonDecoder/JDJsonDecoder.h
@@ -27,7 +27,7 @@
 
 #define JDJsonDecoderErrorDomain @"JDJsonDecoderError"
 
-typedef NS_ENUM(int, JDJsonDecoderError) {
+typedef NS_ENUM(NSUInteger, JDJsonDecoderError) {
     JDJsonDecoderErrorNotDictionary
 };
